Add files via upload
Though the disease is currently spreading around the globe COVID - 19 is also known as Coronavirus. It is solely damaging the health however conjointly the economy of the world as well.

Leveraging today's upgrading technology and the internet for the healthier planet, I have created the Power BI report to highlight the impact of COVID - 19 around the globe and for my country India.

It's our duty to consider this matter on a serious note and follow the guidelines requested by our honorable Prime Minister Narendra Modi punctually.

It's time to come together and fight against this virus phenomenally. For the first time in life, the winner will be the one who will stay put. Let's consider this as a race that will save us and show the world that we are together in this.

Let's unite self-quarantine and fight against the pandemic of the corona.

 Dataset Compiled by 

1. https://www.covid19india.org/

2. https://github.com/CSSEGISandData/COVID-19

3. https://www.worldometers.info/coronavirus/